requirements.txt
pandas
boto3
snowflake-connector-python
python-dotenv
requests